Do not exit with an error on when too low MSI
